[Fix] 수정 사항 반영
