+++ shared/core/presentation/build.gradle.kts
@@ -1,0 +1,57 @@
+plugins {
+    alias(libs.plugins.kotlinMultiplatform)
+    alias(libs.plugins.androidLibrary)
+}
+
+kotlin {
+    androidTarget()
+
+    val xcfName = "shared:core:presentationKit"
+
+    iosX64 {
+        binaries.framework {
+            baseName = xcfName
+        }
+    }
+
+    iosArm64 {
+        binaries.framework {
+            baseName = xcfName
+        }
+    }
+
+    iosSimulatorArm64 {
+        binaries.framework {
+            baseName = xcfName
+        }
+    }
+
+    jvm("desktop")
+
+    sourceSets {
+        commonMain {
+            dependencies {
+            }
+        }
+
+        androidMain {
+            dependencies {
+            }
+        }
+
+        iosMain {
+            dependencies {
+            }
+        }
+    }
+}
+
+android {
+    namespace = "com.kus.core.presentation"
+
+    compileSdk = libs.versions.android.compileSdk.get().toInt()
+
+    defaultConfig {
+        minSdk = libs.versions.android.minSdk.get().toInt()
+    }
+}
