+++ shared/core/presentation/src/androidMain/kotlin/com/kus/core/presentation/Platform.android.kt
@@ -1,0 +1,3 @@
+package com.kus.core.presentation
+
+actual fun platform() = "Android"
