+++ shared/core/presentation/src/commonMain/kotlin/com/kus/core/presentation/Platform.kt
@@ -1,0 +1,3 @@
+package com.kus.core.presentation
+
+expect fun platform(): String
