+++ shared/core/presentation/src/commonMain/kotlin/com/kus/core/presentation/UiError.kt
@@ -1,0 +1,5 @@
+sealed interface UiError {
+    data class Message(val value: String) : UiError
+    data object Network : UiError
+    data object Unauthorized : UiError
+}
