+++ shared/core/presentation/src/commonMain/kotlin/com/kus/core/presentation/UiState.kt
@@ -1,0 +1,6 @@
+sealed interface UiState<out T> {
+    data object Idle : UiState<Nothing>
+    data object Loading : UiState<Nothing>
+    data class Success<T>(val data: T) : UiState<T>
+    data class Failure(val error: UiError) : UiState<Nothing>
+}
