+++ shared/core/presentation/src/desktopMain/kotlin/com/kus/core/presentation/Platform.desktop.kt
@@ -1,0 +1,3 @@
+package com.kus.core.presentation
+
+actual fun platform(): String = "Desktop"
