+++ shared/core/presentation/src/iosMain/kotlin/com/kus/core/presentation/Platform.ios.kt
@@ -1,0 +1,3 @@
+package com.kus.core.presentation
+
+actual fun platform() = "iOS"
