+++ shared/feature/tier/src/commonMain/kotlin/com/kus/feature/tier/TierResultKeys.kt
@@ -1,0 +1,5 @@
+package com.kus.feature.tier
+
+object TierKeys {
+    const val RESULT_FILTER = "tier_result_json"
+}
