+++ shared/feature/tier/src/commonMain/kotlin/com/kus/feature/tier/ui/TierTab.kt
@@ -1,0 +1,6 @@
+package com.kus.feature.tier.ui
+
+enum class TierTab(val title: String) {
+    LIST("티어표"),
+    MAP("지도"),
+}
