+++ shared/feature/tier/src/commonMain/kotlin/com/kus/feature/tier/ui/list/InfiniteScrollEffect.kt
@@ -1,0 +1,26 @@
+package com.kus.feature.tier.ui.list
+
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.snapshotFlow
+
+@Composable
+fun InfiniteScrollEffect(
+    listState: LazyListState,
+    buffer: Int = 4,
+    onLoadMore: () -> Unit,
+) {
+    LaunchedEffect(listState) {
+        snapshotFlow {
+            val layoutInfo = listState.layoutInfo
+            val total = layoutInfo.totalItemsCount
+            val lastVisible = layoutInfo.visibleItemsInfo.lastOrNull()?.index ?: 0
+            total to lastVisible
+        }.collect { (total, lastVisible) ->
+            if (total > 0 && lastVisible >= total - 1 - buffer) {
+                onLoadMore()
+            }
+        }
+    }
+}
